refactor(core): move DSD pipeline to core/audio_format
Step 7 — the DSD container parser, 1-bit-to-PCM decimating FIR, and
DSF/DFF metadata reader (`crate::audio::dsd::*`) migrate as-is to
`waveflow_core::audio_format::dsd::*`. Zero Tauri coupling, zero
audio-engine glue: the module was always self-contained.

What moves:

- `audio_format::dsd::parser` (461 LOC) — DSF (Sony chunk) + DFF
  (Philips IFF/FORM) container layouts, sample rate / channels /
  duration extraction.
- `audio_format::dsd::pcm` (404 LOC) — Blackman-Harris FIR
  decimator that brings the 1-bit bitstream down to 88.2 / 96 kHz
  PCM the rest of the engine already understands. 17 unit tests
  travel with the code.
- `audio_format::dsd::metadata` (370 LOC) — DSF carries an
  embedded ID3v2 footer chunk; this reader hands it to the `id3`
  crate without a full file roundtrip. DFF parses its native
  DIIN / COMT / CMNT atoms inline.

`crates/core/Cargo.toml` picks up `id3 = "1"` for the metadata
side. The DSD module's name as a top-level core entry is
`audio_format` rather than `audio` to mark the boundary cleanly:
the real-time `cpal` / `rtrb` engine stays in `crates/app` per the
Phase 1.a split rules — what lives under `audio_format` is just
format-level processing that's safe to run inside an axum handler
on the future server.

Call-site updates (only two):

- `commands/scan.rs::extract_dsd_file` — imports flip from
  `crate::audio::dsd::*` to `waveflow_core::audio_format::dsd::*`.
- `audio/crossfade.rs` — same flip for the playback decode path.
- `audio/mod.rs` drops its `pub mod dsd;` declaration since the
  directory moved away.

`cargo check --workspace --all-targets` clean; `cargo test
--workspace` 111 passed / 0 failed (66 app + 45 core — DSD's 17
tests + the 8 scanner tests from step 6 + the 20 smart-playlist
tests from 6.c now compile against core).

Refs #128.
