+++ src-tauri/crates/app/src/audio/mod.rs
@@ -17,7 +17,6 @@
 pub mod analytics;
 pub mod crossfade;
 pub mod decoder;
-pub mod dsd;
 pub mod engine;
 pub mod eq;
 pub mod output;
